Initial import of files, not fully working yet
This is the initial import of apptainer definition files for LEPP.
Things are not quite finished yet, but I prefer to have a safety net
while hammering out a first working version.

The idea is to have a Makefile which builds two apptainers:

- one apptainer just contains a version of ROOT (likely to be rebuilt
  rarely)
- one apptainer contains Geant 4 with LEPP/UCN specific patches (to be
  taken off a github repository in the near future), and installs ROOT
  and a few utilities alongside it; the idea is that this container is
  updated a bit more often

The Makefile knows about these dependencies, and uses that to build
locally a working apptainer for people.

TODO: Also import some basic documentation that people can get started
with, and improve as they see fit.