use templating to set Ubuntu version and ROOT branch to build in Makefile
Modify root-ubuntu24.04.def, and rename to root-ubuntu.def to use
templating for the Ubuntu version and the ROOT branch to build. Adapt
Makefile to follow suit, so the Ubuntu version and ROOT branch are now
set at the top of Makefile.

The idea behind this change is to make moving to a new version
relatively hassle-free.